+++ Assets/02_Script/Enemy/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    void TakeDamage(int amount);
+    void TakeDamage(float amount);
 }
